+++ R/tools.R
@@ -484,6 +484,7 @@ get_nrow <- function(varnames, exclude = 0, ncol = 2)  {
 #' Get fig.height for RMarkdown-junk using explore_all()
 #'
 #' @param data A dataset
+#' @param n Weights variable for count data
 #' @param target Target variable
 #' @param nvar Number of variables to plot
 #' @param ncol Number of columns (default = 2)
@@ -495,16 +496,18 @@ get_nrow <- function(varnames, exclude = 0, ncol = 2)  {
 #' total_fig_height(nvar = 5)
 #' @export
 
-total_fig_height <- function(data, target, nvar = NA, ncol = 2, size = 3)  {
+total_fig_height <- function(data, n, target, nvar = NA, ncol = 2, size = 3)  {
 
   if (!is.na(nvar)) {
     n_var <- nvar
-    n <- n_var - ifelse(missing(target), 0, 1)
   } else {
     n_var <- ncol(data)
-    n <- n_var - ifelse(missing(target), 0, 1)
   }
-  result <- ceiling(n / ncol) * size
+
+  n_var <- n_var - ifelse(missing(target), 0, 1)
+  n_var <- n_var - ifelse(missing(n), 0, 1)
+
+  result <- ceiling(n_var / ncol) * size
   result
 }