+++ NEWS.md
@@ -1,11 +1,13 @@
-## explore 0.6.0 (2020-08-14)
+## explore 0.6.0 (2020-09-16)
 
 * new Vignette "explore_penguins"
 * new Vignette "explore_titanic" (count data)
 * explore_count(): plot count() output
+* add default parameter n for count data: 
+  explore(), explore_all, explore_tree(), 
+  describe(), describe_cat, describe_num(), describe_tbl(),
+  total_fig_height() 
 * explore_tree(): default value for minsplit = 10% of obs
-* explore(): add default parameter n for count data
-* explore_tree(): add default parameter n for count data
 * explore_cor(): use geom_point for small datasets
 * explore_shiny(): use browseURL() with parameter browser=NULL
 * describe_tbl(): add observations containing NA